refactor: move BPINN solvers into NeuralPDEBPINNExt package extension
- Move advancedHMC_MCMC.jl, BPINN_ode.jl and PDE_BPINN.jl into ext/bpinn/
  loaded via NeuralPDEBPINNExt; AdvancedHMC, MCMCChains and LogDensityProblems
  become weak dependencies.
- Keep BNNODE / BPINNsolution / BPINNstats structs and vector_to_parameters
  in main src/bpinn_types.jl so the public surface and exports are unchanged;
  only the methods (constructor, solve, ahmc_bayesian_pinn_ode/pde,
  LogTargetDensity / PDELogTargetDensity helpers) live in the extension.
- Add stub functions for ahmc_bayesian_pinn_ode/pde in main with docstrings
  pointing users at the extension load.
- Update Project.toml [weakdeps]/[extensions]/[extras]/[targets] and
  docs/Project.toml so the docs build and the test target still resolve
  AdvancedHMC, MCMCChains and LogDensityProblems.
