+++ docs/Project.toml
@@ -12,8 +12,10 @@ Integrals = "de52edbc-65ea-441a-8357-d3a637375a31"
 IntervalSets = "8197267c-284f-5f27-9208-e0e47529a953"
 LineSearches = "d3d80556-e9d4-5f37-9878-2ab0fcc64255"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+LogDensityProblems = "6fdf6af0-433a-55f7-b3ed-c6c6e0b8df7c"
 Lux = "b2108857-7c20-44ae-9111-449ecde12c47"
 LuxCUDA = "d0bbae9a-e099-4d5b-a835-1c6931763bda"
+MCMCChains = "c7f686f2-ff18-58e9-bc7b-31028e88f75d"
 ModelingToolkit = "961ee093-0014-501f-94e3-6117800e7a78"
 MonteCarloMeasurements = "0987c9cc-fe09-11e8-30f0-b96dd679fdca"
 NeuralOperators = "ea5c82af-86e5-48da-8ee1-382d6ad7af4b"
@@ -42,8 +44,10 @@ Flux = "0.14.17, 0.15, 0.16"
 Integrals = "4, 5.1"
 IntervalSets = "0.7"
 LineSearches = "7.2"
+LogDensityProblems = "2"
 Lux = "1"
 LuxCUDA = "0.3.2"
+MCMCChains = "7"
 ModelingToolkit = "11"
 MonteCarloMeasurements = "1"
 NeuralOperators = "0.5, 0.6"
